Extract waveform viewport state
